Clean up onboarding and settings
