+++ app/src/main/java/xyz/block/gosling/MainActivity.kt
@@ -15,22 +15,13 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.Badge
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
@@ -46,7 +37,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import xyz.block.gosling.ui.theme.GoslingTheme
-import xyz.block.gosling.ui.theme.LocalGoslingColors
 
 class MainActivity : ComponentActivity() {
     companion object {
@@ -147,7 +137,7 @@ fun MainContent(
     var showSettings by remember { mutableStateOf(false) }
 
     if (showSetup) {
-        SetupWizard(
+        Onboarding(
             onSetupComplete = { showSetup = false },
             modifier = modifier,
             settingsManager = settingsManager,
@@ -203,71 +193,6 @@ fun MainContent(
     }
 }
 
-@Composable
-fun MainScreen(modifier: Modifier = Modifier) {
-    val context = LocalContext.current
-    val goslingColors = LocalGoslingColors.current
-    val settingsManager = remember { SettingsManager(context) }
-
-    var isAccessibilityEnabled by remember { mutableStateOf(settingsManager.isAccessibilityEnabled) }
-
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        Text(
-            text = "Gosling Assistant Setup",
-            style = MaterialTheme.typography.headlineMedium
-        )
-
-        // Accessibility Permission Card
-        Card(
-            modifier = Modifier.fillMaxWidth(),
-            shape = RoundedCornerShape(8.dp),
-            colors = CardDefaults.cardColors(
-                containerColor = goslingColors.inputBackground
-            )
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                horizontalAlignment = Alignment.Start
-            ) {
-                Text(
-                    text = "Accessibility Permissions",
-                    style = MaterialTheme.typography.titleMedium
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-                Text(
-                    text = "Allow Gosling to interact with other apps",
-                    style = MaterialTheme.typography.bodyMedium
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-                Button(
-                    onClick = {
-                        openAccessibilitySettings(context)
-                        isAccessibilityEnabled = settingsManager.isAccessibilityEnabled
-                    },
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = if (isAccessibilityEnabled) goslingColors.secondaryButton else goslingColors.primaryBackground
-                    )
-                ) {
-                    Text(if (isAccessibilityEnabled) "Accessibility Enabled" else "Enable Accessibility")
-                }
-            }
-        }
-
-        // Gosling UI
-        if (isAccessibilityEnabled) {
-            GoslingUI(context = context)
-        }
-    }
-}
-
 fun checkAccessibilityPermission(context: Context): Boolean {
     val enabledServices = Settings.Secure.getString(
         context.contentResolver,
@@ -276,9 +201,4 @@ fun checkAccessibilityPermission(context: Context): Boolean {
     val isEnabled = enabledServices?.contains(context.packageName) == true
     Log.d("Gosling", "Accessibility check: $enabledServices, enabled: $isEnabled")
     return isEnabled
-}
-
-fun openAccessibilitySettings(context: Context) {
-    val intent = Intent(Settings.ACTION_ACCESSIBILITY_SETTINGS)
-    context.startActivity(intent)
 }
