+++ app/src/main/java/xyz/block/gosling/Settings.kt
@@ -1,0 +1,280 @@
+package xyz.block.gosling
+
+import android.provider.Settings
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExposedDropdownMenuBox
+import androidx.compose.material3.ExposedDropdownMenuDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Switch
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.unit.dp
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SettingsScreen(
+    settingsManager: SettingsManager,
+    onBack: () -> Unit,
+    openAccessibilitySettings: () -> Unit,
+    isAccessibilityEnabled: Boolean
+) {
+    val context = LocalContext.current
+    var isAssistantEnabled by remember { mutableStateOf(false) }
+    var llmModel by remember { mutableStateOf(settingsManager.llmModel) }
+    var apiKey by remember { mutableStateOf(settingsManager.apiKey) }
+    var shouldProcessNotifications by remember { mutableStateOf(settingsManager.shouldProcessNotifications) }
+    var showResetDialog by remember { mutableStateOf(false) }
+    var expanded by remember { mutableStateOf(false) }
+
+    // Function to check assistant status
+    fun checkAssistantStatus() {
+        val settingSecure = Settings.Secure.getString(
+            context.contentResolver,
+            "assistant"
+        )
+        isAssistantEnabled = settingSecure?.contains(context.packageName) == true
+    }
+
+    // Check on initial launch
+    LaunchedEffect(Unit) {
+        checkAssistantStatus()
+    }
+
+    // Check when app regains focus
+    DisposableEffect(Unit) {
+        val activity = context as? android.app.Activity
+        val lifecycleObserver = object : android.app.Application.ActivityLifecycleCallbacks {
+            override fun onActivityResumed(activity: android.app.Activity) {
+                if (activity == context) {
+                    checkAssistantStatus()
+                }
+            }
+            
+            override fun onActivityPaused(activity: android.app.Activity) {}
+            override fun onActivityStarted(activity: android.app.Activity) {}
+            override fun onActivityDestroyed(activity: android.app.Activity) {}
+            override fun onActivitySaveInstanceState(activity: android.app.Activity, outState: android.os.Bundle) {}
+            override fun onActivityStopped(activity: android.app.Activity) {}
+            override fun onActivityCreated(activity: android.app.Activity, savedInstanceState: android.os.Bundle?) {}
+        }
+
+        activity?.application?.registerActivityLifecycleCallbacks(lifecycleObserver)
+
+        onDispose {
+            activity?.application?.unregisterActivityLifecycleCallbacks(lifecycleObserver)
+        }
+    }
+
+    val models = listOf(
+        "gpt-4o" to "GPT-4 Optimized",
+        "o3-mini" to "O3 Mini",
+        "o3-small" to "O3 Small",
+        "o3-medium" to "O3 Medium",
+        "o3-large" to "O3 Large"
+    )
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text("Settings") },
+                navigationIcon = {
+                    IconButton(onClick = onBack) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+                    }
+                }
+            )
+        }
+    ) { paddingValues ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(paddingValues)
+                .padding(horizontal = 24.dp),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.SpaceBetween
+        ) {
+            Column(
+                modifier = Modifier.fillMaxWidth(),
+                verticalArrangement = Arrangement.spacedBy(32.dp)
+            ) {
+                Column(
+                    modifier = Modifier.fillMaxWidth(),
+                    verticalArrangement = Arrangement.spacedBy(24.dp)
+                ) {
+                    Column(
+                        modifier = Modifier.fillMaxWidth(),
+                        verticalArrangement = Arrangement.spacedBy(8.dp)
+                    ) {
+                        Text(text = "LLM Model")
+                        ExposedDropdownMenuBox(
+                            expanded = expanded,
+                            onExpandedChange = { expanded = it }
+                        ) {
+                            OutlinedTextField(
+                                value = models.find { it.first == llmModel }?.second ?: llmModel,
+                                onValueChange = {},
+                                readOnly = true,
+                                modifier = Modifier
+                                    .fillMaxWidth()
+                                    .menuAnchor(),
+                                trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded) }
+                            )
+
+                            ExposedDropdownMenu(
+                                expanded = expanded,
+                                onDismissRequest = { expanded = false }
+                            ) {
+                                models.forEach { (modelId, displayName) ->
+                                    DropdownMenuItem(
+                                        text = { Text(displayName) },
+                                        onClick = {
+                                            llmModel = modelId
+                                            settingsManager.llmModel = modelId
+                                            expanded = false
+                                        }
+                                    )
+                                }
+                            }
+                        }
+                    }
+
+                    // API Key
+                    Column(
+                        modifier = Modifier.fillMaxWidth(),
+                        verticalArrangement = Arrangement.spacedBy(8.dp)
+                    ) {
+                        Text(text = "API Key")
+                        OutlinedTextField(
+                            value = apiKey,
+                            onValueChange = {
+                                apiKey = it
+                                settingsManager.apiKey = it
+                            },
+                            modifier = Modifier.fillMaxWidth(),
+                            visualTransformation = PasswordVisualTransformation(),
+                            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password)
+                        )
+                    }
+                }
+
+                // Accessibility/Notifications Section
+                Column(
+                    modifier = Modifier.fillMaxWidth(),
+                    verticalArrangement = Arrangement.spacedBy(16.dp)
+                ) {
+                    if (isAccessibilityEnabled) {
+                        Text(
+                            text = "Notification Processing",
+                            style = MaterialTheme.typography.titleLarge,
+                        )
+                        Text(
+                            text = "Allow Gosling to analyze and respond to notifications from other apps",
+                            style = MaterialTheme.typography.bodyMedium,
+                            color = MaterialTheme.colorScheme.onSurfaceVariant
+                        )
+                        Row(
+                            modifier = Modifier.fillMaxWidth(),
+                            horizontalArrangement = Arrangement.SpaceBetween,
+                            verticalAlignment = Alignment.CenterVertically
+                        ) {
+                            Text(
+                                text = "Process notifications",
+                                style = MaterialTheme.typography.bodyLarge
+                            )
+                            Switch(
+                                checked = shouldProcessNotifications,
+                                onCheckedChange = { 
+                                    shouldProcessNotifications = it
+                                    settingsManager.shouldProcessNotifications = it
+                                }
+                            )
+                        }
+                    } else {
+                        Text(
+                            text = "Accessibility Permissions",
+                            style = MaterialTheme.typography.titleLarge,
+                        )
+                        Text(
+                            text = "Gosling needs accessibility permissions to interact with other apps and help you with tasks.",
+                            style = MaterialTheme.typography.bodyMedium,
+                            color = MaterialTheme.colorScheme.onSurfaceVariant
+                        )
+                        Button(
+                            onClick = openAccessibilitySettings,
+                            modifier = Modifier.fillMaxWidth(),
+                            enabled = true,
+                        ) {
+                            Text("Enable Accessibility")
+                        }
+                    }
+                }
+            }
+
+            // Clear Settings Button at bottom
+            Button(
+                onClick = { showResetDialog = true },
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 32.dp),
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = MaterialTheme.colorScheme.error
+                )
+            ) {
+                Text("Clear saved settings")
+            }
+        }
+    }
+
+    if (showResetDialog) {
+        AlertDialog(
+            onDismissRequest = { showResetDialog = false },
+            title = { Text("Reset Setup") },
+            text = { Text("This will reset all settings and show the setup wizard again. Are you sure?") },
+            confirmButton = {
+                TextButton(
+                    onClick = {
+                        settingsManager.isFirstTime = true
+                        showResetDialog = false
+                    }
+                ) {
+                    Text("Reset")
+                }
+            },
+            dismissButton = {
+                TextButton(onClick = { showResetDialog = false }) {
+                    Text("Cancel")
+                }
+            }
+        )
+    }
+}
